requirements.txt
# General & Environment
pandas
numpy
python-dotenv
typing-extensions

# Person 1: Agent Reasoning & Orchestration
langchain
langchain-core
langchain-community
langchain-openai
openai
pydantic
tiktoken

# Person 2: Quantitative Analysis & Stats
scipy
statsmodels
openpyxl

# Person 2: Qualitative Analysis (NLP)
sentence-transformers
spacy
textblob
langdetect
scikit-learn

# Person 2: Visualization (For Goal 3 Heatmaps/Journey Maps)
matplotlib
seaborn

# Person 2: Vector DB & RAG
chromadb
faiss-cpu

# Person 3: UI & Integration
streamlit
requests
fastapi
uvicorn